Move MemberPage to separate file
