+++ scraper.rb
@@ -4,7 +4,6 @@
 
 require 'pry'
 require 'require_all'
-require 'scraped'
 require 'scraperwiki'
 
 require_rel 'lib'
@@ -12,42 +11,6 @@ require_rel 'lib'
 # require 'open-uri/cached'
 # OpenURI::Cache.cache_path = '.cache'
 require 'scraped_page_archive/open-uri'
-
-class MemberPage < Scraped::HTML
-  decorator Scraped::Response::Decorator::AbsoluteUrls
-
-  field :id do
-    url.to_s.split('/').last
-  end
-
-  field :name do
-    box.css('h1').text.split(' - ').first.sub('Hon. ', '').sub(' MP', '').tidy
-  end
-
-  field :faction do
-    f = box.xpath('.//strong[contains(.,"Parliamentary Group")]/..//img/@title').text
-    return 'Partit Nazzjonlista' if f == 'PN'
-    f
-  end
-
-  field :email do
-    box.css('a[href*="mailto:"]/@href').text.split('mailto:').drop(1).first
-  end
-
-  field :image do
-    box.css('img/@src').first.text
-  end
-
-  field :source do
-    url.to_s
-  end
-
-  private
-
-  def box
-    noko.css('div.column2')
-  end
-end
 
 def scrape(h)
   url, klass = h.to_a.first
